requirements.txt
bcrypt==4.0.1
cffi==1.15.1
colorama==0.4.6
cryptography==40.0.2
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.1
paramiko==3.1.0
pluggy==1.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyNaCl==1.5.0
pytest==7.3.1
python-dotenv==1.0.0
soupsieve==2.4.1
sshtunnel==0.4.0
tomli==2.0.1
